==> notebooks/00_initial_experiments.ipynb <==
Notes:
- Users will likely not want to be recommended an anime they have already watched
    - Additional filtering will need to be done during inference
- Users might be more emotionally invested in a subsequent season of an anime
    - Additional weight for subsequent season of a series? (may be future work)
- Users likely prefer more current anime
    - Have model favor recent releases
    - https://static.googleusercontent.com/media/research.google.com/en//pubs/archive/45530.pdf, section 3.3

# Dataset sources: 
# https://www.kaggle.com/datasets/CooperUnion/anime-recommendations-database
# https://www.kaggle.com/datasets/crazygump/myanimelist-scrappind-a-decade-of-anime

import pandas as pd

df_anime_rating = pd.read_csv('../data/01_raw/anime.csv')
df_anime_meta = pd.read_csv('../data/01_raw/MAL-all-from-winter1917-to-fall2022.csv')
df_user_rating = pd.read_csv('../data/01_raw/rating.csv')

print(df_anime_rating.shape)
display(df_anime_rating.head())

print(df_anime_meta.shape)
display(df_anime_meta.head())

print(df_user_rating.shape)
display(df_user_rating.head())

df_user_rating['anime_id']

- rating of -1 means a user has watched a show, but has not rated it (https://www.kaggle.com/datasets/CooperUnion/anime-recommendations-database/discussion/253814)
- Suggestion: use https://jikan.moe/ OR https://github.com/abhinavk99/jikanpy jinkan API to extract studio, producer ()information? (note: rate limited to 60 request per minute)
- Use https://github.com/QasimK/mal-scraper to scrape?
- Scrapy (https://scrapy.org/) can be used to easily extract synopsis, reviews and image links
- https://www.kaggle.com/datasets/crazygump/myanimelist-scrappind-a-decade-of-anime/data can be merged on MAL_id 
- one hot encode genre
- VAE encoding for features?
- NCF for recommender system? https://medium.com/data-science-in-your-pocket/recommendation-systems-using-neural-collaborative-filtering-ncf-explained-with-codes-21a97e48a2f7

df_anime_meta.rename(columns={"MAL_id":"anime_id"}, inplace=True)
df_anime = df_anime_rating.merge(df_anime_meta, on='anime_id')

# Check for missing values

df_anime_rating.info()

df_user_rating['rating'].value_counts()

df_anime_meta.info()

for col in df_anime_meta.columns:
    print(df_anime_meta[col].value_counts())
