==> notebooks/01_data_download.ipynb <==
# Contents
1. Download kaggle datasets
2. Scrape following from MAL websites:
    - Synopsis
    - Background
    - Voice actors (TBC)
    - Image links

Data preprocessing intended for next ipynb

## 1. Download kaggle datasets

import kaggle
import pandas as pd


ds_path = "../data/01_raw"

kaggle.api.authenticate()

ds1_kagglename = "CooperUnion/anime-recommendations-database"
ds2_kagglename = "crazygump/myanimelist-scrappind-a-decade-of-anime"

kaggle.api.dataset_download_files(ds1_kagglename, path=ds_path, unzip=True)
kaggle.api.dataset_download_files(ds2_kagglename, path=ds_path, unzip=True)

## 2. Scrape data from MAL Websites

import pandas as pd 

# Gathering unique anime_id values from ratings dataset:
ratings_path = "../data/01_raw/rating.csv"
df_ratings = pd.read_csv(ratings_path)

# df_unique_anime = df_ratings["anime_id"].unique()
df_unique_anime = df_ratings.drop_duplicates(subset=['anime_id'])
print(len(df_unique_anime))

partition_size = int(len(df_unique_anime)/4)
df_unique_anime_damien = df_unique_anime[:partition_size]
df_unique_anime_leroy = df_unique_anime[partition_size:2*partition_size]
df_unique_anime_rosamund = df_unique_anime[2*partition_size:3*partition_size]
df_unique_anime_kenneth = df_unique_anime[3*partition_size:4*partition_size]

print(df_unique_anime_damien.info())
print(df_unique_anime_leroy.info())
print(df_unique_anime_rosamund.info())
print(df_unique_anime_kenneth.info())

# Export unique anime IDs for separate scrapy scraping
df_unique_anime_damien.to_csv("../data/01_raw/rating_damien.csv")
df_unique_anime_leroy.to_csv("../data/01_raw/rating_leroy.csv")
df_unique_anime_rosamund.to_csv("../data/01_raw/rating_rosamund.csv")
df_unique_anime_kenneth.to_csv("../data/01_raw/rating_kenneth.csv")
