==> notebooks/02_data_merge_&_rescrape.ipynb <==
import pandas as pd
import numpy as np

df_consolidated = pd.read_csv('../data/01_raw/anime_scrapy_consolidated.csv', sep="|")
df_scraped = pd.read_csv('../data/01_raw/anime_scrapy.csv', sep="|")
df_rating = pd.read_csv('../data/01_raw/rating.csv')

df_consolidated = pd.concat([df_consolidated, df_scraped])

print(f'# of scraped unique anime_ids: {len(df_consolidated["anime_id"].unique())}')
print(f'# of unique anime ids: {len(df_rating["anime_id"].unique())}')
print(f'# of remaining anime ids to scrape:{len(df_rating["anime_id"].unique())-len(df_consolidated["anime_id"].unique())}')

# Export remaining anime_id list
missing_anime = np.setdiff1d(df_rating['anime_id'],df_consolidated['anime_id'])
print(f'Length of missing_anime list: {len(missing_anime)}')
np.savetxt("../data/01_raw/missing_anime.csv", missing_anime, header="anime_id", fmt="%i", comments="")

target_cols = [
        "anime_id", 
        "synopsis", 
        "image_url", 
        "rating", 
        "va_list", 
        "staff_list",
        "recommended_review_count",
        "mixedfeelings_review_count",
        "notrecommended_review_count"
        ]

df_consolidated = df_consolidated[target_cols]
df_consolidated.to_csv('../data/01_raw/anime_scrapy_consolidated.csv', sep="|", index=False)
