==> notebooks/03a_primary_dataset.ipynb <==
import pandas as pd
import numpy as np

df_ds1 = pd.read_csv("../data/01_raw/anime.csv")
df_ds2 = pd.read_csv("../data/01_raw/MAL-all-from-winter1917-to-fall2022.csv")
df_scraped = pd.read_csv("../data/01_raw/anime_scrapy_consolidated.csv", sep="|")
df_missing = pd.read_csv("../data/01_raw/missing_anime.csv")


df_ds1.head()

df_ds2.head()

df_list = [df_ds1, df_ds2, df_scraped, df_missing]

for df in df_list:
    print(df.columns)

# Merging kaggle datasets 1 and 2
df_left = df_ds1[['anime_id', 'name', 'type', 'episodes', 'rating', 'members']]
df_right = df_ds2[['MAL_id', 'studio', 'release-season', 'release-year', 'release-date', 'source-material', 'genres', 'themes', 'demographics']]
df = pd.merge(df_left, df_right , how="left", left_on="anime_id", right_on="MAL_id")
df = df.drop(columns="MAL_id")
df.rename(columns={"rating":"mal_score"}, inplace=True)
df.head()

# Merging dataset with scraped dataset
df_left = df
df_right = df_scraped
df = pd.merge(df_left, df_right , how="left", on="anime_id")
pd.set_option('display.max_columns', None)
df.head()

# Removing missing anime from dataset and rating
df_final = df
df_final = df_final[~df['anime_id'].isin(df_missing['anime_id'])]
df_final = df_final.drop_duplicates(subset="anime_id", keep="first")
df_final.head()

df_rating = pd.read_csv('../data/01_raw/rating.csv')
df_final_rating = df_rating[~df_rating["anime_id"].isin(df_missing['anime_id'])]

# Exporting dataset to 03_primary
anime_output_path = "../data/03_primary/anime_final.csv"
df_final.to_csv(anime_output_path, sep="|", index=False)

rating_output_path = "../data/03_primary/rating_final.csv"
df_final_rating.to_csv(rating_output_path, sep="|", index=False)
