==> notebooks/04a_embded_image.ipynb <==
# Image download

import pandas as pd
import numpy as np
from PIL import Image
import requests
from tqdm import tqdm
import os

df_anime = pd.read_csv('../data/03_primary/anime_final.csv', sep='|')

df_anime.head(3)

print(df_anime['image_url'].isna().sum())
df_anime['image_url'] = df_anime['image_url'].fillna("No URL")
print(df_anime['image_url'].isna().sum())

if os.path.isfile('../data/05_model_input/missing_anime.csv'):
    df_anime = pd.read_csv('../data/05_model_input/missing_anime.csv', sep='|')
    for i in tqdm(range(len(df_anime))):
        url = df_anime['image_url'][i]
        if url == "No URL":
            pass
        else:
            img = Image.open(requests.get(url, stream=True).raw)
            # Export image
            filepath = f"../data/05_model_input/images/{df_anime['anime_id'][i]}.jpg"
            img.save(filepath)


loaded_img_list = os.listdir('../data/05_model_input/images')
loaded_img_df = pd.DataFrame(loaded_img_list, columns=['loaded_img_path'])
loaded_img_df['anime_id']  = loaded_img_df.apply(lambda x: os.path.splitext(x['loaded_img_path'])[0], axis=1).astype(int)
display(loaded_img_df.head())

print(f'# of loaded anime images: {len(loaded_img_list)}')
print(f'Total # of anime images:s: {len(df_anime["image_url"])}')
print(f'# of remaining anime ids to scrape:{len(df_anime["image_url"])-len(loaded_img_list)}')


# Export remaining anime_id list
# missing_anime = np.setdiff1d(df_anime['anime_id'], loaded_img_df['anime_id']).astype(int)
missing_anime = df_anime[~(df_anime['anime_id'].isin(loaded_img_df['anime_id']))][['anime_id','image_url']]
print(missing_anime)
print(f'Length of missing_anime list: {len(missing_anime)}')
missing_anime.to_csv('../data/05_model_input/missing_anime.csv',sep='|', index=False)

# Image tensor extraction

import pandas as pd
import numpy as np
from tqdm import tqdm

import tensorflow as tf
from tensorflow.keras.applications import VGG16
from tensorflow.keras.preprocessing import image
from tensorflow.keras.applications.vgg16 import preprocess_input

df_anime = pd.read_csv('../data/03_primary/anime_final.csv',sep="|")
df_anime['image_url'] = df_anime['image_url'].fillna('None')
df_anime.head(3)

img_tensors = []
model = VGG16(weights='imagenet', include_top=False)

for i in tqdm(range(len(df_anime))):
    if df_anime['image_url'][i] == "None":
        img_tensors.append(None)
    else:
        img_path = f"../data/05_model_input/images/{df_anime['anime_id'][i]}.jpg"
        img = image.load_img(img_path, target_size=(224, 224))
        x = image.img_to_array(img)
        x = np.expand_dims(x, axis=0)
        x = preprocess_input(x)
        features = model.predict(x, verbose=0)
        features_flat = features.flatten()
        img_tensors.append(features_flat)

df_anime['image_features'] = img_tensors

df_anime.to_csv('../data/05_model_input/anime_final_[img].csv',sep="|")

df_anime.head(3)
