==> notebooks/04b_embed_synopsis.ipynb <==
Created by: Rosamund

## Overview

NOTE: 
Additional packages required: 
- pip install -U sentence-transformers==3.0.0

To avoid TqdmWarning: IProgress not found:
- pip install --upgrade jupyter ipywidgets


A Sentence Transformer is a type of natural language processing model designed specifically to produce meaningful and useful sentence embeddings. Sentence embeddings are fixed-length numerical representations that capture the semantic meaning of a sentence.

Reference: https://github.com/VishalS-HK/product-recommendation-system-BERT/blob/main/Product_Recommendation_System_BERT.ipynb

## Import Libraries

import pandas as pd
import numpy as np
from sentence_transformers import SentenceTransformer
import tqdm as tqdm

## Extract data

anime_final_df = pd.read_csv('../data/anime_final.csv', sep="|")

print(f'No. of rows: {anime_final_df.shape[0]:,}')
print(f'No. of columns: {anime_final_df.shape[1]:,}')

anime_final_df.head(3)

Print out the first synopsis

anime_final_df.iloc[0]['synopsis']

anime_final_df.iloc[100]['synopsis']

Preprocessing:
At the end of every synopsis there is '[Written by MAL Rewrite]'. We want to remove this

anime_final_df['synopsis'] = anime_final_df['synopsis'].str.replace(r'\s*\[Written by MAL Rewrite\]\s*','',regex=True)

Let's inspect if that phrase is still there 

anime_final_df.iloc[0]['synopsis']

anime_final_df.iloc[100]['synopsis']

Type conversion

anime_final_df[anime_final_df['synopsis'].isna()]

## Sentence Transformers

Instantiate Bert sentence transformer

model = SentenceTransformer('bert-base-nli-mean-tokens')

Generate embeddings

# testing on first 5
synopsis_list = anime_final_df['synopsis'].iloc[:5].tolist()
sentence_embeddings = model.encode(synopsis_list, show_progress_bar=True)

# Note: for .encode(), there is an optional arugment: normalize_embeddings (bool, optional) – Whether to 
# normalize returned vectors to have length 1. In that case, the faster dot-product (util.dot_score) instead 
# of cosine similarity can be used. Defaults to False.

sentence_embeddings

Issue: Could not complete the embedding because there are NaN values for the 'synopsis column'.
